requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
PyJWT>=2.10.0

